refactor: moves all `type`s into seperate file
