+++ src/apps/vendor-finance/common/vendor-finance.pool-v2.contract-position-fetcher.ts
@@ -1,14 +1,11 @@
 import { Inject } from '@nestjs/common';
-
 import { APP_TOOLKIT, IAppToolkit } from '~app-toolkit/app-toolkit.interface';
-import { PositionTemplate } from '~app-toolkit/decorators/position-template.decorator';
 import { DefaultDataProps } from '~position/display.interface';
 import { ContractPositionTemplatePositionFetcher } from '~position/template/contract-position.template.position-fetcher';
 import {
   GetDefinitionsParams,
   DefaultContractPositionDefinition,
   GetTokenDefinitionsParams,
-  UnderlyingTokenDefinition,
   GetDisplayPropsParams,
   GetTokenBalancesParams,
   GetDataPropsParams,
