+++ src/apps/vendor-finance/common/vendor-finance.pool.contract-position-fetcher.ts
@@ -22,23 +22,7 @@ import { VendorFinanceContractFactory, VendorFinancePool } from '../contracts';
 
 import { borrowerInfosQuery, VendorBorrowerGraphResponse } from './getBorrowerInfosQuery';
 import { LENDING_POOLS_QUERY, VendorLendingPoolsGraphResponse } from './getLendingPoolsQuery';
-
-export type VendorFinancePoolDefinition = {
-  address: string;
-  deployer: string;
-  mintRatio: string;
-  colToken: string;
-  lendToken: string;
-  expiry: string;
-  feeRate: string;
-  lendBalance: string;
-  totalBorrowed: string;
-};
-
-type VendorFinancePoolDataProps = DefaultDataProps & {
-  deployer: string;
-  totalDeposited: number;
-};
+import { VendorFinancePoolDataProps, VendorFinancePoolDefinition } from './vendor-finance.pool.types';
 
 export abstract class VendorFinancePoolContractPositionFetcher extends ContractPositionTemplatePositionFetcher<VendorFinancePool> {
   abstract subgraphUrl: string;
